pyproject.toml
[tool.poetry]
name = "pathfinder"
version = "0.1.0"
description = ""
authors = ["Stuart Fyfe <stuart@fyfe.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.2"
pyyaml = "6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
